backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
